Bb XML Parsed
The Blackboard XML parsing seems more or less accomplished, and we can handle files up to 250MB without a problem (or, at least, our biggest test file is 215MB and it worked fine). I've diagrammed out the conversion from Blackboard objects to Canvas objects, and now we're down to the fiddly bits.

(N.B. I have rejiggered the .gitignore file so that I can maintain a file with a plaintext copy of the API key and matching instance URL without committing it to Git accidentally.)

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>