User Class
Identify a user based on a unique class in which they are enrolled and modify the menu to reflect those permissions. (Currently based on Faculty Resources -- which is removed from the Courses menu for Faculty and moved to Resources, but roles could be added.)

Added core of Lion Hub/SMS to the menu.

Signed-off-by: Seth Battis <seth@battis.net>