Independent JQuery UI menubar
So. Appending content to the Canvas menus dynamically is [no longer working](https://github.com/smtech/canvashack-plugin-global-navigation-menu/issues/8) (although it will be interesting to see if the sidebars in the new UI are more flexible in this regard).

I've ripped out the styling that makes the menus appear to be native Canvas global navigation menus, and replaced it with a JQuery UI menu. (Hint: the big trick to making a JQuery UI menu appear as a horizontal menubar is:

```CSS
.ui-menu > .ui-menu-item {
    display: inline-block;
    float: left;
    clear: none; /* without this, float left is meaningless */
    position: relative;
    width: auto !important;
}
```

I'm going to need to restructure my actual menu items, but at least now the menus themselves are being added and are operational.